+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+1.0a2 (2017-2-20)
++++++++++++++++++
+
+* Add search options to admin.
+* Export current options command.
+
 1.0a1 (2017-2-20)
 +++++++++++++++++